{{ cookiecutter.project_slug }}/requirements-development.txt
-r requirements.txt

autoflake >= 2.3.1, < 3.0.0
black >= 24.4.2, < 25.0.0
coverage >= 7.6.12, < 8.0.0
{%- if cookiecutter.enable_channels == "y" %}
daphne >= 4.2.1, < 5.0.0  # Required for testing channels: <https://github.com/django/channels/issues/1942>
{%- endif %}
flake8 >= 7.0.0, < 8.0.0
isort >= 5.13.2, < 6.0.0
model-bakery >= 1.17.0, < 1.18.0
pytest >= 8.2.0, < 9.0.0
{%- if cookiecutter.enable_channels == "y" %}
pytest-asyncio >= 0.24, < 1
{%- endif %}
pytest-django >= 4.8.0, < 5.0.0
pytest-env >= 1.1.3, < 2.0.0
sqlparse >= 0.5.0, < 0.6.0
